requirements.txt
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
dnspython==1.16.0
exchangelib==3.1.1
idna==2.8
isodate==0.6.0
lxml==4.5.0
ntlm-auth==1.4.0
oauthlib==3.1.0
pycparser==2.19
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
six==1.14.0
tzlocal==2.0.0
urllib3==1.25.8
